implemented easy to use text system
to create a new overworld character, once you have the object representing it drag the "interactable" and "dialogue loader" scripts in the scripts folder onto your object. Then on the right in the inspector it will ask you for a json, which you will drag into the json field. A character sprite which you can also drag. and a sound effect which you also drag. After that, thats it! thats all you have to do! once you're in proximity of this character in the overworld and press 'e' then it will let you interact with it, and the text will come up. if you have any questions ask me and i can make it work but this is as streamlined as I think it should be